requirements.txt
fastapi
uvicorn
jinja2
langchain
langchain-google-genai
google-generativeai
faiss-cpu
langchain-community
langchain-cli
pandas